+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # latest
-Using version [1.1](#11-2025-06-12)
+Using version [1.2](#12-2025-06-20)
 
 # Major version 1
-Using version [1.1](#11-2025-06-12)
+Using version [1.2](#12-2025-06-20)
+
+## 1.2 (2025-06-20)
+**Changes**
+- Whisper model will now be unloaded after 10 minutes of queue inactivity. Model will be reloaded when a new item is added to the queue.
+    - Ideally, this should resolve a bug that occurs when the model is loaded for days at a time.
+    - Model will still be loaded at startup. This is used to ensure that the model is downloaded before any other action occurs.
+- Increased queue size warning log to 10 from 5.
 
 ## 1.1 (2025-06-12)
 **Changes**